pyproject.toml
[project]
name = "bus-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.6.5",
    "celery>=5.5.3",
    "fastapi>=0.115.12",
    "fastapi-sso>=0.18.0",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "mariadb>=1.1.13",
    "mysql-connector-python>=9.3.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pwdlib[argon2]>=0.2.1",
    "pydantic[email]>=2.11.2",
    "pyjwt>=2.10.1",
    "pytest>=8.4.2",
    "python-decouple>=3.8",
    "python-multipart>=0.0.20",
    "ruff>=0.11.3",
    "sentry-sdk[fastapi]>=2.41.0",
    "sqlalchemy>=2.0.40",
    "sqlalchemy-utils>=0.41.2",
    "sqlmodel>=0.0.24",
    "uvicorn[standard]>=0.34.0",
]
